First commit, this is still a work in progress.
I just updated the security in this API so it using spring-security-web/spring-security-core version 6.3.3 with spring-boot-starter-web and spring-boot-starter-oauth2-resource-server version 3.3.3.  and yes, I am still working out the kinks with the updated new security. ;-)
But this should give anyone a good sample of my code style when it comes to Java Rest APIs.